move group logic
